net: ethernet: mtk_eth_soc: handle dma buffer size soc specific
The mainline MTK ethernet driver suffers long time from rarly but
annoying tx queue timeouts. We think that this is caused by fixed
dma sizes hardcoded for all SoCs.

We suspect this problem arises from a low level of free TX DMADs,
the TX Ring alomost full.

The transmit timeout is caused by the Tx queue not waking up. The
Tx queue stops when the free counter is less than ring->thres, and
it will wake up once the free counter is greater than ring->thres.
If the CPU is too late to wake up the Tx queues, it may cause a
transmit timeout.
Therefore, we increased the TX and RX DMADs to improve this error
situation.

Use the dma-size implementation from SDK in a per SoC manner. In
difference to SDK we have no RSS feature yet, so all RX/TX sizes
should be raised from 512 to 2048 byte except fqdma on mt7988 to
avoid the tx timeout issue.

Fixes: 656e705243fd ("net-next: mediatek: add support for MT7623 ethernet")
Suggested-by: Daniel Golle <daniel@makrotopia.org>
Signed-off-by: Frank Wunderlich <frank-w@public-files.de>
---
based on SDK:

https://git01.mediatek.com/plugins/gitiles/openwrt/feeds/mtk-openwrt-feeds/+/fac194d6253d339e15c651c052b532a449a04d6e

v3:
- rephrase commit message with some information from mtk
- change the TX/RX DMA Size from 512 to 2048 for all platforms.
  When rss gets upstream TX Size should be increases to 4096 and
  the RX Size decreased to 1024 for MT798[1/6/8].
- drop fq_dma_size from rt5350/mt7628 as this does not have QDMA
v2:
- fix unused variable 'addr' in 32bit build

---

mtk-info (not to be upstreamed):

We've encountered a transmit queue timeout issue on the MT79888 and
have identified it as being related to the RSS feature.

Since RSS is enabled, there are 4 Rx Rings, with each containing 2048
DMADs, totaling 8192 for Rx. In contrast, the Tx Ring has only 2048
DMADs. Tx DMADs will be consumed rapidly during a 10G LAN to 10G WAN
forwarding test, subsequently causing the transmit queue to stop.
Therefore, we reduced the number of Rx DMADs for each ring to balance
both Tx and Rx DMADs, which resolves this issue.

Yes, we believe that changing the TX/RX DMA Size from 512 to 2048 can
improve the transmit timeout issue.

Additionally, the patch you provided contains two changes:
1. Change the TX/RX DMA Size from 512 to 2048 for all platforms.
2. Change the TX Size to 4096 and the RX Size to 1024 for MT798[1/6/8].
I think that currently, only the first part needs to be submitted. Once
RSS is upstreamed in the future, we can then put the second part into
net-next.

Therefore, you only need to change the TX/RX DMA Size from 512 to 2048
for all platforms, and change the FQ DMA Size to 4096 for MT7988, while
keeping the others at 2048.

About the trigger tx timeout, we believe that if the available TX DMAD
becomes very low, there is a chance to trigger a transmit timeout.