gcc: disable '-Warray-bounds' for gcc-9
'-Warray-bounds' is already disabled for gcc-10+. Now that we've merged
bitmap_{read,write), I see the following error when building the kernel
with gcc-9.4 (Ubuntu 20.04.4 LTS) for x86_64 allmodconfig:

/home/yury/work/linux/drivers/pinctrl/pinctrl-cy8c95x0.c: In function ‘cy8c95x0_read_regs_mask.isra.0’:
/home/yury/work/linux/include/linux/bitmap.h:756:18: error: array subscript [1, 288230376151711744] is outside array bounds of ‘long unsigned int[1]’ [-Werror=array-bounds]
  756 |  value_high = map[index + 1] & BITMAP_LAST_WORD_MASK(start + nbits);
      |               ~~~^~~~~~~~~~~

The immediate reason is that the commit b44759705f7dc ("bitmap: make
bitmap_{get,set}_value8() use bitmap_{read,write}(") switched the
bitmap_get_value8() to an alias of bitmap_read(); the same for 'set'.

Now; the code that triggers Warray-bounds, calls the function like this:

DECLARE_BITMAP(tval, MAX_LINE); // 64-bit map: unsigned long tval[1]

read_val |= bitmap_get_value8(tval, i * BANK_SZ) & ~bits;

bitmap_read() is implemented in a way that it may conditionally
dereference a pointer beyond the boundary like this:

        unsigned long space = BITS_PER_LONG - offset;

        if (space >= nbits)
                return (map[index] >> offset) & BITMAP_LAST_WORD_MASK(nbits);

        value_low = map[index] & BITMAP_FIRST_WORD_MASK(start);
        value_high = map[index + 1] & BITMAP_LAST_WORD_MASK(start + nbits);
        return (value_low >> offset) | (value_high << space);

But in case of bitmap_get_value8(), it's clearly impossible to violate the
boundary with new implementation, as soon as the function is passed with
correct parameters.

So, this is clearly a false-positive.

It has already been spotted in this thread:

https://lore.kernel.org/linux-kernel//457b9cbb-9a5f-47ef-9eac-3e4f135d6a96@intel.com/T/#mc833ceac1b05d7156a864dcfe21435d01d77a0d7

And I suggested to use the OPTIMIZER_HIDE_VAR() there. But now that I
see the warning disabled for gcc-10+, it looks better to just extend
it to gcc-9.

Signed-off-by: Yury Norov <yury.norov@gmail.com>