net: phy: as21xxx: implement read workaround for C45 read
This PHY have lots of problems with MDIO read operation. We somehow
workaround this with using C45 operation for pretty much everything but
this is not enough. The reference code for this PHY makes a write to an
unused PHY to workaround this read problem. This was also confirmed by
Aeonsemi.

Various test were made to try to workaround this ins alternative way
than the random write.

One effective solution was to limit the write only to BMSR. And also
write to BMSR is safe since they are only read only registers.

This is only done for read operation as write operation doesn't suffer
from this problem.

Worth to mention that when multiple Aeonsemi PHY are mounted, the
workaround doesn't work if we write to another Aeonsemi PHY.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>